remove cistem-specific import ctf protocol
