+++ cistem/protocols.conf
@@ -19,9 +19,6 @@ Protocols SPA = [
 	]}
 	]
 Tomography = [
-	{"tag": "section", "text": "Imports", "icon": "bookmark.png", "children": [
-	    {"tag": "protocol", "value": "CistemProtTsImportCtf", "text": "default"}
-	]},
 	{"tag": "section", "text": "Tilt-series", "children": [
 		{"tag": "protocol_group", "text": "Tilt-series preprocess", "openItem": "False", "children": []},
 		{"tag": "protocol_group", "text": "CTF", "openItem": "False", "children": [
